requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
mysqlclient==2.0.1
pkg-resources==0.0.0
python-dotenv==0.14.0
requests==2.23.0
SQLAlchemy==1.3.18
urllib3==1.25.10
websockets==8.1
yarl==1.4.2
